+++ pypeit/par/pypeitpar.py
@@ -913,7 +913,7 @@ class Coadd1DPar(ParSet):
         defaults['wave_method'] = 'linear'
         dtypes['wave_method'] = str
         descr['wave_method'] = "Method used to construct wavelength grid for coadding spectra. The routine that creates " \
-                               "the wavelength is coadd1d.get_wave_grid. The options are:" \
+                               "the wavelength is :func:`~pypeit.core.wavecal.wvutils.get_wave_grid`. The options are:" \
                                " "\
                                "'iref' -- Use the first wavelength array" \
                                "'velocity' -- Grid is uniform in velocity" \
@@ -1360,6 +1360,7 @@ class SensFuncPar(ParSet):
 
         # Initialize the other used specifications for this parameter set
         defaults = OrderedDict.fromkeys(pars.keys())
+        options = OrderedDict.fromkeys(pars.keys())
         dtypes = OrderedDict.fromkeys(pars.keys())
         descr = OrderedDict.fromkeys(pars.keys())
 
@@ -1391,6 +1392,7 @@ class SensFuncPar(ParSet):
 
         defaults['algorithm'] = 'UVIS'
         dtypes['algorithm'] = str
+        options['algorithm'] = SensFuncPar.valid_algorithms()
         descr['algorithm'] = "Specify the algorithm for computing the sensitivity function. The options are: " \
                              " (1) UVIS = Should be used for data with lambda < 7000A." \
                              "No detailed model of telluric absorption but corrects for atmospheric extinction." \
@@ -1435,9 +1437,10 @@ class SensFuncPar(ParSet):
         super(SensFuncPar, self).__init__(list(pars.keys()),
                                           values=list(pars.values()),
                                           defaults=list(defaults.values()),
+                                          options=list(options.values()),
                                           dtypes=list(dtypes.values()),
                                           descr=list(descr.values()))
-        self.validate()
+#        self.validate()
 
     @classmethod
     def from_dict(cls, cfg):
@@ -1455,13 +1458,12 @@ class SensFuncPar(ParSet):
             kwargs[pk] = cfg[pk] if pk in k else None
         return cls(**kwargs)
 
-    def validate(self):
+    @staticmethod
+    def valid_algorithms():
         """
-        Check the parameters are valid for the provided method.
+        Return the valid sensitivity algorithms.
         """
-        if not ((self.data['algorithm'] == 'IR') or  (self.data['algorithm'] == 'UVIS')):
-            raise ValueError('algorithm must be set to either  "IR" or "UVIS"')
-        # JFH add other checks?
+        return ['UVIS', 'IR']
 
 
 class SensfuncUVISPar(ParSet):